pyproject.toml
[project]
name = "docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "google-generativeai>=0.8.4",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
    "trafilatura>=1.8.0",
]
